Bank Application
I learned how to make a GUI with the java bank application demo. In this GUI i used 3 action event listeners which represented the checking account, savings account and exit button. In both the checking and saving account buttons extended classes were used to subnet the actions of making a user account to this application